+++ src/main/java/com/ll/exam/sbb/question/repository/QuestionRepository.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface QuestionRepository extends JpaRepository<Question, Integer>, RepositoryUtil {
+public interface QuestionRepository extends JpaRepository<Question, Long>, RepositoryUtil {
     Question findBySubject(String subject);
 
     Question findBySubjectAndContent(String subject, String content);
