+++ src/main/java/com/ll/exam/sbb/question/service/QuestionService.java
@@ -32,7 +32,7 @@ public class QuestionService {
         return this.questionRepository.findAll(pageable);
     }
 
-    public Question getQuestion(int id) {
+    public Question getQuestion(Long id) {
 
         return questionRepository.findById(id)
                 .orElseThrow(() -> new DataNotFoundException("no %d question not found,".formatted(id)));
